saas_rbac/requirements.txt

Django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
redis>=5.0
celery>=5.3
python-dotenv>=1.0
drf-spectacular>=0.26
gunicorn>=21.2